Добавила трейт с функциями которые выводят разницу во времени
