+++ frontend/models/Tasks.php
@@ -34,6 +34,7 @@ use Yii;
  */
 class Tasks extends \yii\db\ActiveRecord
 {
+    use TimeCreationToCurrentTrait;
     /**
      * {@inheritdoc}
      */
@@ -150,44 +151,5 @@ class Tasks extends \yii\db\ActiveRecord
         return $this->hasOne(Cities::class, ['id' => 'city_id']);
     }
 
-    /**
-     * @return \DateInterval
-     */
-    private function getTimeCreationToCurrent(): \DateInterval
-    {
-        $current = new \DateTime();
-        return $current->diff(new \DateTime($this->created));
-    }
 
-    /**
-     * @return int
-     */
-    public function getYear(): int
-    {
-        return $this->getTimeCreationToCurrent()->y;
-    }
-
-    /**
-     * @return int
-     */
-    public function getMonth(): int
-    {
-        return $this->getTimeCreationToCurrent()->m;
-    }
-
-    /**
-     * @return int
-     */
-    public function getDay(): int
-    {
-        return $this->getTimeCreationToCurrent()->d;
-    }
-
-    /**
-     * @return int
-     */
-    public function getHour(): int
-    {
-        return $this->getTimeCreationToCurrent()->h;
-    }
 }
